+++ client/voting-game/src/components/sectionComponent.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Route, Routes } from "react-router-dom";
 import { getRandomJoke } from '../services/jokeService';
 import { JokeType } from '../types/jokeType';
-import Joke from './joke';
+import Joke from './Joke/joke';
 import NextBtn from './nextBtn';
 import UpdateComponent from './updateComponent';
